refactor: display 요소를 컴포넌트로 분리 #9
display를 컴포넌트로 분리
