+++ src/styles/pages/Home.css
@@ -7,23 +7,4 @@
 
   width: 100%;
   height: 100%;
-
-}
-
-.display {
-  position: absolute;
-  top: 15%;
-  display: flex;
-
-  width: 90%;
-  height: 80%;
-
-  justify-content: center;
-  align-items: center;
-
-
-  border: solid 2px white;
-  border-radius: 5%;
-
-  overflow: hidden;
 }
